requirements.txt
fastapi[standard-no-fastapi-cloud-cli]
google-genai
